+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## applin Ruby Gem Changes
+- v0.7.0
+    - Support ApplinIos 0.32.0.
+    - Remove on_user_error_poll action and make it a parameter of the rpc action.
+    - Add `modal` action.
+    - Add `aspect_ratio` param to `choose_photo` and `take_photo` actions.
 - v0.6.0
     - Support ApplinIos v0.25.0.
     - Add `poll_delay_ms` to checkbox and textfield.